@@ -1,21 +1,15 @@
 {
   "homepage": ".",
-  "name": "react-typescript-starter-pack",
+  "name": "react-todo-app",
   "version": "0.1.0",
   "dependencies": {
-    "@testing-library/jest-dom": "^5.12.0",
-    "@testing-library/react": "^11.2.6",
-    "@testing-library/user-event": "^12.8.3",
-    "@types/jest": "^26.0.23",
-    "@types/node": "^12.20.10",
-    "@types/react": "^17.0.4",
-    "@types/react-dom": "^17.0.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-scripts": "4.0.3",
-    "typescript": "^4.2.4",
-    "typescript-eslint-parser": "^22.0.0",
-    "web-vitals": "^1.1.1"
+    "@fortawesome/fontawesome-free": "^6.1.2",
+    "bulma": "^0.9.4",
+    "classnames": "^2.3.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-scripts": "5.0.1",
+    "react-transition-group": "^4.4.5"
   },
   "scripts": {
     "lint": "eslint ./src --fix",
@@ -45,9 +39,26 @@
     ]
   },
   "devDependencies": {
-    "@mate-academy/eslint-config-react-typescript": "^0.0.4",
-    "eslint": "^7.25.0",
-    "gh-pages": "^3.1.0",
-    "node-sass": "^5.0.0"
+    "@cypress/react": "^5.12.4",
+    "@cypress/webpack-dev-server": "^1.8.4",
+    "@mate-academy/cypress-tools": "^1.0.4",
+    "@mate-academy/eslint-config-react-typescript": "^1.0.11",
+    "@mate-academy/scripts": "^1.2.8",
+    "@mate-academy/students-ts-config": "*",
+    "@mate-academy/stylelint-config": "*",
+    "@types/node": "^17.0.45",
+    "@types/react": "^18.0.15",
+    "@types/react-dom": "^18.0.6",
+    "@types/react-transition-group": "^4.4.5",
+    "cypress": "^9.5.3",
+    "eslint": "^7.32.0",
+    "eslint-plugin-cypress": "^2.11.2",
+    "gh-pages": "^3.2.3",
+    "mochawesome": "^7.1.3",
+    "mochawesome-merge": "^4.2.0",
+    "mochawesome-report-generator": "^6.2.0",
+    "node-sass": "^7.0.1",
+    "stylelint": "^13.13.1",
+    "typescript": "^4.6.3"
   }
 }
